requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.0
pydantic==2.11.4
email-validator==2.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
redis==5.0.8
rq==1.16.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
